make new source folder on Jongro branch and imported copies of two files
