+++ .env.development
@@ -1,5 +1,5 @@
 GENERATE_SOURCEMAP = false
 REACT_APP_RTSP_IP=211.34.80.54
 REACT_APP_MODEL_PATH=glb/Jongro
-REACT_APP_PROGRESS_TIMEOUT=100000
+REACT_APP_PROGRESS_TIMEOUT=10000
 REACT_APP_VIDEO_URL=/video/drone1.mp4
