+++ package.json
@@ -12,6 +12,7 @@
     "@react-three/drei": "9.77.0",
     "@react-three/fiber": "8.12.0",
     "@types/three": "0.149.0",
+    "gsap": "^3.12.2",
     "leva": "0.9.34",
     "react": "18.2.0",
     "react-dom": "18.2.0",
