+++ public/source/output.json
@@ -1,0 +1,7 @@
+[
+    {
+        "filename": "crack_4.glb",
+        "mesh_name": "Tile_32002",
+        "material_name": "Tile_32_0.001"
+    }
+]
